+++ packages/database/src/index.ts
@@ -13,6 +13,7 @@ export { createStudiosRepository } from "./repositories/studios-repository";
 export { createUsersRepository } from "./repositories/users-repository";
 export { createInvitesRepository } from "./repositories/invites-repository";
 export { createProjectsRepository } from "./repositories/projects-repository";
+export { createEpicsRepository } from "./repositories/epics-repository";
 export { createGamesRepository } from "./repositories/games-repository";
 export { createKnowledgeDocumentsRepository } from "./repositories/knowledge-documents-repository";
 export { createSubmissionsRepository } from "./repositories/submissions-repository";